Better packaging in data directory
