+++ app/src/androidTest/java/com/google/samples/apps/sunflower/compose/plantdetail/PlantDetailComposeTest.kt
@@ -27,9 +27,7 @@ import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
 import androidx.core.net.toUri
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.google.samples.apps.sunflower.compose.plantdetail.PlantDetails
-import com.google.samples.apps.sunflower.compose.plantdetail.PlantDetailsCallbacks
-import com.google.samples.apps.sunflower.data.Plant
+import com.google.samples.apps.sunflower.data.models.Plant
 import com.google.samples.apps.sunflower.test.R
 import org.junit.Rule
 import org.junit.Test
