+++ app/src/androidTest/java/com/google/samples/apps/sunflower/data/PlantDaoTest.kt
@@ -20,6 +20,9 @@ import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.room.Room
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
+import com.google.samples.apps.sunflower.data.local.dao.PlantDao
+import com.google.samples.apps.sunflower.data.local.database.AppDatabase
+import com.google.samples.apps.sunflower.data.models.Plant
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
 import org.hamcrest.MatcherAssert.assertThat
