+++ app/src/main/java/com/google/samples/apps/sunflower/api/UnsplashService.kt
@@ -17,7 +17,7 @@
 package com.google.samples.apps.sunflower.api
 
 import com.google.samples.apps.sunflower.BuildConfig
-import com.google.samples.apps.sunflower.data.UnsplashSearchResponse
+import com.google.samples.apps.sunflower.data.models.UnsplashSearchResponse
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import okhttp3.logging.HttpLoggingInterceptor.Level
