+++ app/src/main/java/com/google/samples/apps/sunflower/compose/gallery/GalleryScreen.kt
@@ -50,9 +50,9 @@ import androidx.paging.compose.collectAsLazyPagingItems
 import androidx.paging.compose.itemKey
 import com.google.samples.apps.sunflower.R
 import com.google.samples.apps.sunflower.compose.plantlist.PhotoListItem
-import com.google.samples.apps.sunflower.data.UnsplashPhoto
-import com.google.samples.apps.sunflower.data.UnsplashPhotoUrls
-import com.google.samples.apps.sunflower.data.UnsplashUser
+import com.google.samples.apps.sunflower.data.models.UnsplashPhoto
+import com.google.samples.apps.sunflower.data.models.UnsplashPhotoUrls
+import com.google.samples.apps.sunflower.data.models.UnsplashUser
 import com.google.samples.apps.sunflower.viewmodels.GalleryViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
