+++ app/src/main/java/com/google/samples/apps/sunflower/compose/plantlist/PlantListItemView.kt
@@ -35,8 +35,8 @@ import androidx.compose.ui.text.style.TextAlign
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
 import com.google.samples.apps.sunflower.R
-import com.google.samples.apps.sunflower.data.Plant
-import com.google.samples.apps.sunflower.data.UnsplashPhoto
+import com.google.samples.apps.sunflower.data.models.Plant
+import com.google.samples.apps.sunflower.data.models.UnsplashPhoto
 
 @Composable
 fun PlantListItem(plant: Plant, onClick: () -> Unit) {
