+++ app/src/main/java/com/google/samples/apps/sunflower/data/UnsplashPagingSource.kt
@@ -19,6 +19,7 @@ package com.google.samples.apps.sunflower.data
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
 import com.google.samples.apps.sunflower.api.UnsplashService
+import com.google.samples.apps.sunflower.data.models.UnsplashPhoto
 
 private const val UNSPLASH_STARTING_PAGE_INDEX = 1
