+++ app/src/main/java/com/google/samples/apps/sunflower/data/local/dao/GardenPlantingDao.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Google LLC
+ * Copyright 2024 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,15 @@
  * limitations under the License.
  */
 
-package com.google.samples.apps.sunflower.data
+package com.google.samples.apps.sunflower.data.local.dao
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Transaction
+import com.google.samples.apps.sunflower.data.models.GardenPlanting
+import com.google.samples.apps.sunflower.data.models.PlantAndGardenPlantings
 import kotlinx.coroutines.flow.Flow
 
 /**
