+++ app/src/main/java/com/google/samples/apps/sunflower/data/models/Plant.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.samples.apps.sunflower.data
+package com.google.samples.apps.sunflower.data.models
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
