+++ app/src/main/java/com/google/samples/apps/sunflower/data/models/UnsplashPhoto.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Google LLC
+ * Copyright 2024 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,18 +14,21 @@
  * limitations under the License.
  */
 
-package com.google.samples.apps.sunflower.data
+package com.google.samples.apps.sunflower.data.models
 
 import com.google.gson.annotations.SerializedName
+import com.google.samples.apps.sunflower.data.models.UnsplashPhotoUrls
+import com.google.samples.apps.sunflower.data.models.UnsplashUser
 
 /**
- * Data class that represents a photo search response from Unsplash.
+ * Data class that represents a photo from Unsplash.
  *
  * Not all of the fields returned from the API are represented here; only the ones used in this
  * project are listed below. For a full list of fields, consult the API documentation
- * [here](https://unsplash.com/documentation#search-photos).
+ * [here](https://unsplash.com/documentation#get-a-photo).
  */
-data class UnsplashSearchResponse(
-    @field:SerializedName("results") val results: List<UnsplashPhoto>,
-    @field:SerializedName("total_pages") val totalPages: Int
+data class UnsplashPhoto(
+    @field:SerializedName("id") val id: String,
+    @field:SerializedName("urls") val urls: UnsplashPhotoUrls,
+    @field:SerializedName("user") val user: UnsplashUser
 )
