+++ app/src/main/java/com/google/samples/apps/sunflower/data/models/UnsplashUser.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.samples.apps.sunflower.data
+package com.google.samples.apps.sunflower.data.models
 
 import com.google.gson.annotations.SerializedName
