+++ app/src/main/java/com/google/samples/apps/sunflower/data/repositories/UnsplashRepository.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Google LLC
+ * Copyright 2024 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,14 @@
  * limitations under the License.
  */
 
-package com.google.samples.apps.sunflower.data
+package com.google.samples.apps.sunflower.data.repositories
 
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import com.google.samples.apps.sunflower.api.UnsplashService
+import com.google.samples.apps.sunflower.data.UnsplashPagingSource
+import com.google.samples.apps.sunflower.data.models.UnsplashPhoto
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
